fail safe

+++ src/Listeners/Client/dms.ts
@@ -7,6 +7,7 @@ import { Logger } from "winston";
 import { main, sec, third, fourth, a1, split } from "../../lib/Utils";
 import alexa from "alexa-bot-api";
 const ai = new alexa();
+let spokenusers = [];
 export default class message extends Listener {
 	client: Client;
 	public constructor(client: Client) {
@@ -19,13 +20,15 @@ export default class message extends Listener {
 	}
 
 	async exec(message: Message): Promise<void | Logger> {
-		let spokenusers = [];
-
-		if (message.channel.id == "807302538558308355") {
-			if (spokenusers.includes(message.author.id)) return;
-			spokenusers.push(message.author.id);
-			const reg = /^<@!?304054669372817419> happy birthday <:clownmonke:807704911973187585>$/i;
-			if (message.content.search(reg) == -1) message.delete();
+		this.client.gp.ensure("on", true);
+		const on = this.client.gp.get("on");
+		if (on == true) {
+			if (message.channel.id == "807302538558308355") {
+				if (spokenusers.includes(message.author.id)) return;
+				spokenusers.push(message.author.id);
+				const reg = /^<@!?304054669372817419> happy birthday <:clownmonke:807704911973187585>$/i;
+				if (message.content.search(reg) == -1) message.delete();
+			}
 		}
 
 		if (message.channel.id == "823935750168117312") {
